requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.94.1
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==43.0.1
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.18.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.115.0
filelock==3.25.2
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.10.2
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.19.0
numpy==2.2.6
packaging==24.2
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
plotly==5.24.0
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.9
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.20.0
PyJWT==2.9.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rich==13.9.4
rpds-py==0.30.0
scikit-learn==1.5.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smmap==5.0.3
sniffio==1.3.1
starlette==0.38.6
streamlit==1.38.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
